% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{addMMPAS2PACds}
\alias{addMMPAS2PACds}
\title{Add animal's polyA signal (PAS) to a PACdataset.}
\usage{
addMMPAS2PACds(pacds, bsgenome, from = -100, to = 10, label = "PAS_")
}
\arguments{
\item{pacds}{A PACdataset object.}

\item{bsgenome}{chrmosome fasta files, an object of BSgenome or FaFile, see faFromPACds().}

\item{from}{to specify the range near PACs, PAC is the 0 position.
e.g., from=-50, to=-1 to subset 50 nt (PAC is the 0 or 51st position, upstream 50nt of PAC), see faFromPACds().}

\item{to}{similar to from.}

\item{label}{A character string specifying the output column prefix. The function adds one or two columns, including `label_gram` and `dist`.
The _gram column gives the gram that is closest to the PAC.
_dist is the start position of a gram to a PAC.}
}
\value{
A PACdataset with columns such as `label_gram` and `dist` added.
}
\description{
addMMPAS2PACds adds animal's polyA signal (PAS) to a PACdataset.
This function considers the priority of different PAS and search animal PAS, AATAAA 100nt -> ATTAAA 100nt -> TGTA 100nt -> others 50nt.
}
\examples{
\dontrun{
pdAPA.mut=addMMPAS2PACds(pdAPA.mut, bsgenome,
                        from=-100, to=10, label='PAS_')
}
}
\seealso{
Other APAds functions:
\code{\link[=APAds2GLMds]{APAds2GLMds()}},
\code{\link[=QAPA2GLMds]{QAPA2GLMds()}},
\code{\link[=addM6A2PACds]{addM6A2PACds()}},
\code{\link[=addPAS2PACds]{addPAS2PACds()}},
\code{\link[=get3UTRAPApd]{get3UTRAPApd()}},
\code{\link[=get3UTRAPApdDE]{get3UTRAPApdDE()}},
\code{\link[=mergePACdsCoords]{mergePACdsCoords()}},
\code{\link[=mergePACdsRanges]{mergePACdsRanges()}}
}
\concept{APAds functions}
