% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{addPAS2PACds}
\alias{addPAS2PACds}
\title{Add polyA signal (PAS) to a PACdataset.}
\usage{
addPAS2PACds(
  pacds,
  bsgenome,
  grams,
  from,
  to,
  priority = 1,
  label = "PAS_",
  chrCheck = TRUE
)
}
\arguments{
\item{pacds}{A PACdataset object.}

\item{bsgenome}{chrmosome fasta files, an object of BSgenome or FaFile, see faFromPACds().}

\item{grams}{a character vector to specify a gram like AATAAA, or v1 (AATAAA's variants), or multiple grams. grams can be not equal length, like c('AATAAA','AAATTT','CCCT')}

\item{from}{to specify the range near PACs, PAC is the 0 position.
e.g., from=-50, to=-1 to subset 50 nt (PAC is the 0 or 51st position, upstream 50nt of PAC), see faFromPACds().}

\item{to}{similar to from.}

\item{priority}{a numeric vector to set the priority and subgroups of grams if grams has multiple elements, default is NULL.
For example, if grams=c('AATAAA','ATTAAA','AAAAAA','TTTAT), priority=(1,2,3,3), then will first search for AATAAA, if not exists, then for ATTAAA, then the remaining AAAAAA/TTTAT.
If priority=NULL, then will treat all elements in grams as the same group (no priority).}

\item{label}{label A character string specifying the output column prefix. The function adds one or two columns, including `label_gram` and `dist`.
The _gram column gives the gram that is closest to the PAC.
_dist is the start position of a gram to a PAC.}

\item{chrCheck}{if TRUE, then all chr in PACds should be in bsgenome, otherwise will ignore those non-consistent chr rows in PACds.}
}
\value{
A PACdataset with columns such as `label_gram` and `dist` added.
}
\description{
addPAS2PACds adds polyA signal (PAS) to a PACdataset. This function adds PAS_dist/gram/coord columns to PACds@anno.
The first position of coord is the first A in AATAAA; dist is the distance from the last nt (upstream) or first nt (downstream) of AATAAA to the polyA site.
The polyA site position is 0, and the extracted sequence is `from+to+1` nt.
Unlike movAPA's annotateByPAS, in this function, dist has positive and negative values, and will add the coord column, and will change the NA to NOPAS.
}
\examples{
\dontrun{
paAPA.AATAAA=addPAS2PACds(pdAPA, bsgenome, grams='AATAAA',
                         from=-100, to=10, label='AATAAA_')
}
}
\seealso{
Other APAds functions:
\code{\link[=APAds2GLMds]{APAds2GLMds()}},
\code{\link[=QAPA2GLMds]{QAPA2GLMds()}},
\code{\link[=addM6A2PACds]{addM6A2PACds()}},
\code{\link[=addMMPAS2PACds]{addMMPAS2PACds()}},
\code{\link[=get3UTRAPApd]{get3UTRAPApd()}},
\code{\link[=get3UTRAPApdDE]{get3UTRAPApdDE()}},
\code{\link[=mergePACdsCoords]{mergePACdsCoords()}},
\code{\link[=mergePACdsRanges]{mergePACdsRanges()}}
}
\concept{APAds functions}
