% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{addPost2FitModels}
\alias{addPost2FitModels}
\title{Add posterior var to fitted models}
\usage{
addPost2FitModels(models, robust = FALSE)
}
\arguments{
\item{models}{A model list.}

\item{robust}{Default is FALSE. See limma::squeezeVar's robust.}
}
\value{
A model list with updated var.post and df.post.
}
\description{
This function adds posterior var to fitted models by calling limma::squeezeVar.
}
\seealso{
Other QBGLM functions:
\code{\link[=FitModel]{FitModel()}},
\code{\link[=QBGLM]{QBGLM()}},
\code{\link[=QBGLMmultiDs]{QBGLMmultiDs()}},
\code{\link[=fitQuasiGLM]{fitQuasiGLM()}},
\code{\link[=getM6AperGene]{getM6AperGene()}},
\code{\link[=getRUDperGene]{getRUDperGene()}},
\code{\link[=removeFitErrorModels]{removeFitErrorModels()}},
\code{\link[=testBetas]{testBetas()}}
}
\concept{QBGLM functions}
