% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{addQAPARawCounts}
\alias{addQAPARawCounts}
\title{Add raw counts to QAPA results}
\usage{
addQAPARawCounts(qfile, sfs, suffix = "CNT", ofile = NULL)
}
\arguments{
\item{qfile}{A file name of the QAPA APA list.}

\item{sfs}{A vector of quant.sf files of QAPA's outputs which record raw counts. Names of `sfs` should be sample names in the `qfile`.}

\item{suffix}{Default is 'CNT'. Will add columns named `sample.CNT` in the `qfile`.}

\item{ofile}{If not NULL, then output the data.frame to ofile.}
}
\value{
A file name or a data.frame. The function will double check the ID and TPM columns to ensure that each QAPA row receives a unique CNT.
}
\description{
addQAPARawCounts Add raw counts from QAPA's quant.sf files to QAPA's APA list.
}
\examples{
\dontrun{
qfile='sim3_qapa_hg38.txt'
sfs=c('QAPA_saf/qapa_SRR847370_quant.sf', 'QAPA_saf/qapa_SRR847371_quant.sf',
      'QAPA_saf/qapa_SRR847374_quant.sf', 'QAPA_saf/qapa_SRR847375_quant.sf')
names(sfs)=paste0('sample', 1:4)
ofile='sim3_qapa_hg38_CNT.txt'
q=addQAPARawCounts(qfile, sfs, ofile=ofile)
}
}
\seealso{
Other APAdata functions:
\code{\link[=APAtrap2PACds]{APAtrap2PACds()}},
\code{\link[=QAPA2PACds]{QAPA2PACds()}},
\code{\link[=bed2PACds]{bed2PACds()}}
}
\concept{APAdata functions}
