% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{APAds2GLMds}
\alias{APAds2GLMds}
\title{Wrapper function to convert APA's PACdataset to PACdataset for QBGLM.}
\usage{
APAds2GLMds(
  APAds,
  PDminDist = 50,
  PDmaxDist = 5000,
  PDminRatio = 0.05,
  PDfixDistal = FALSE,
  DEpval = 0.05,
  txdb,
  extUTRlen = 2000
)
}
\arguments{
\item{APAds}{A pacdataset of APA sites.}

\item{PDminDist}{Default is 50. Require the distance between  proximal and distal polyA site to be > minDist and < maxDist.}

\item{PDmaxDist}{Default is 1000. Require the distance between  proximal and distal polyA site to be > minDist and < maxDist.}

\item{PDminRatio}{Default is 0.05. If >0 then will filter out polyA sites that do not meet the requirements based on the ratio first, and then filter proximal and distal sites.}

\item{PDfixDistal}{Default is FALSE. True to fix the farthest PA, that is, only look for the proximal site.}

\item{DEpval}{cutoff to filter DE-APA, default is 0.05. If =1, then not filter DE.}

\item{txdb}{An TXDB object for genome annotation.}

\item{extUTRlen}{Length to extend 3'UTR to include more near downstream polyA sites.}
}
\value{
A PACdataset with only proximal and distal, and DE-APA (if DEpval<1) sites.
}
\description{
APAds2GLMds is a wrapper function to convert APA's PACdataset to PACdataset for QBGLM. It annotates polyA sites with TXDB and get proximal and distal sites of 3'UTR-APA genes. Finally, it filters DE-APA by fisher.test of replicates.
}
\examples{
\dontrun{
## first, get a APAds from QAPA results
# add raw counts from .sf files to QAPA list
qfile='M3KD.qapa.txt'
sfs=c('qapa_SRR847370_quant.sf', 'qapa_SRR847371_quant.sf',
      'qapa_SRR847374_quant.sf', 'qapa_SRR847375_quant.sf')
names(sfs)=paste0('sample', 1:4)
q=addQAPARawCounts(qfile, sfs, suffix='CNT', ofile=NULL)
head(q)

# convert QAPA list to PACdataset
apads=QAPA2PACds(q, vcol='CNT')

# change sample names: ctrlM3 and treatM3 and 1/2 for reps
apads=setPACdsSmpInfo(apads,
                smpInfo=cbind(old=paste0('sample', 1:4),
                              new=c('ctrlM31','ctrlM32','treatM31','treatM32'),
                              group=c('ctrlM3','ctrlM3','treatM3','treatM3')))
apads@anno$gene_QAPA=apads@anno$gene
movAPA::summary(apads)

## get GLMds
apads=APAds2GLMds(APAds=apads,
          PDminDist=50, PDmaxDist=5000, PDminRatio=0.05, PDfixDistal=FALSE,
          DEpval=0.05,
          txdb=txdb,
          extUTRlen=2000)
}
}
\seealso{
Other APAds functions:
\code{\link[=QAPA2GLMds]{QAPA2GLMds()}},
\code{\link[=addM6A2PACds]{addM6A2PACds()}},
\code{\link[=addMMPAS2PACds]{addMMPAS2PACds()}},
\code{\link[=addPAS2PACds]{addPAS2PACds()}},
\code{\link[=get3UTRAPApd]{get3UTRAPApd()}},
\code{\link[=get3UTRAPApdDE]{get3UTRAPApdDE()}},
\code{\link[=mergePACdsCoords]{mergePACdsCoords()}},
\code{\link[=mergePACdsRanges]{mergePACdsRanges()}}
}
\concept{APAds functions}
