% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{APAtrap2PACds}
\alias{APAtrap2PACds}
\title{convert APAtrap results to PACdataset}
\usage{
APAtrap2PACds(d)
}
\arguments{
\item{d}{A file name of APAtrap output.}
}
\value{
PACdataset.
}
\description{
APAtrap2PACds convert APAtrap results to PACdataset.
}
\examples{
\dontrun{
apads=APAtrap2PACds('APAtrap.txt')
}
}
\seealso{
Other APAdata functions:
\code{\link[=QAPA2PACds]{QAPA2PACds()}},
\code{\link[=addQAPARawCounts]{addQAPARawCounts()}},
\code{\link[=bed2PACds]{bed2PACds()}}
}
\concept{APAdata functions}
