% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{bed2PACds}
\alias{bed2PACds}
\title{Annotate a BED file and convert it to a PACdataset.}
\usage{
bed2PACds(d, txdb)
}
\arguments{
\item{d}{A BED file name.}

\item{txdb}{a txdb object.}
}
\value{
A PACdataset, with the BED file annotated by ChIPseeker::annotatePeak.
}
\description{
bed2PACds annotates a BED file and converts it to a PACdataset.
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb=TxDb.Hsapiens.UCSC.hg19.knownGene
d='DM/diff_treat_vs_control_c3.0_cond1.bed'
p1.txdb=bed2PACds(d, txdb) #5615 entrezID
p2.gff=bed2PACds(d, txdb) #5585 genesymbol
## #The gene_start/end obtained from the two txdbs may have some differences,
## but the annotations are similar
p1.txdb[p1.txdb@anno$coord==62550816]
p2.gff[p2.gff@anno$coord==62550816]
}
}
\seealso{
Other APAdata functions:
\code{\link[=APAtrap2PACds]{APAtrap2PACds()}},
\code{\link[=QAPA2PACds]{QAPA2PACds()}},
\code{\link[=addQAPARawCounts]{addQAPARawCounts()}}
}
\concept{APAdata functions}
