% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{.getInputNames}
\alias{.getInputNames}
\title{Get m6A Input sample names}
\usage{
.getInputNames(smpCols)
}
\arguments{
\item{smpCols}{Character vector with IP sample names first and Input
sample names second.}
}
\value{
A character vector containing the Input sample names.
}
\description{
Extracts the second half of the sample columns, which should contain
Input samples.
}
