% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{.getIPnames}
\alias{.getIPnames}
\title{Get m6A IP sample names}
\usage{
.getIPnames(smpCols)
}
\arguments{
\item{smpCols}{Character vector with IP sample names first and Input
sample names second.}
}
\value{
A character vector containing the IP sample names.
}
\description{
Extracts the first half of the sample columns, which should contain
m6A IP samples.
}
