% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{.getRdaData}
\alias{.getRdaData}
\title{Read peak information from an RDA file}
\usage{
.getRdaData(rdafile)
}
\arguments{
\item{rdafile}{Path to an RDA file.}
}
\value{
A list containing peak information, library sizes, and sample names.
}
\description{
Read peak information from an RDA file
}
