% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{.m6AExpressDf2PACds}
\alias{.m6AExpressDf2PACds}
\title{Convert an m6A expression data.frame to a PACdataset}
\usage{
.m6AExpressDf2PACds(peakDf, smpCols)
}
\arguments{
\item{peakDf}{A data.frame containing normalized m6A expression values.}

\item{smpCols}{Character vector of sample columns.}
}
\value{
A PACdataset object.
}
\description{
Convert an m6A expression data.frame to a PACdataset
}
\keyword{internal}
