% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{.statPQ}
\alias{.statPQ}
\title{Summarize p-value and q-value results}
\usage{
.statPQ(resMp, pthd = 0.05, qthd = 0.2)
}
\arguments{
\item{resMp}{A data.frame containing p-value columns.}

\item{pthd}{The p-value threshold. Default is 0.05.}

\item{qthd}{The q-value threshold. Default is 0.2.}
}
\value{
This function prints summary statistics and returns invisibly.
}
\description{
Summarize p-value and q-value results
}
