% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{.subsetResMp}
\alias{.subsetResMp}
\title{Subset results using p-value and q-value thresholds}
\usage{
.subsetResMp(resMp, rawPval1 = 1, pval1 = 1, qval1 = 1)
}
\arguments{
\item{resMp}{A data.frame containing result statistics.}

\item{rawPval1}{Raw p-value threshold. Default is 1.}

\item{pval1}{Adjusted or moderated p-value threshold. Default is 1.}

\item{qval1}{Q-value threshold. Default is 1.}
}
\value{
A filtered and sorted data.frame.
}
\description{
Filter a result data.frame using raw p-value, adjusted p-value,
and q-value thresholds.
}
