% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{filterHVPerGene}
\alias{filterHVPerGene}
\title{Get highly variable genes by m6A and RUD scores}
\usage{
filterHVPerGene(
  perGeneData,
  m6Acutoff = -1,
  RUDcutoff = -1,
  totalWeights = -1,
  perWeights = 10,
  nWeights = 6,
  minAvgM6A = 0.1
)
}
\arguments{
\item{perGeneData}{A data frame of per gene data with m6A$ and RUD$ columns, which can be obtained by getRUDperGene and getM6AperGene.}

\item{m6Acutoff}{Cutoff for m6A score; if >0, then do filtering.}

\item{RUDcutoff}{Cutoff for RUD score; if >0, then do filtering.}

\item{totalWeights}{Cutoff for weights (total counts); if >0, then do filtering.}

\item{perWeights}{Cutoff for weights per sample; if >0, then do filtering. Default is 10.}

\item{nWeights}{Number of samples with weights>=perWeights, default is 6.}

\item{minAvgM6A}{Cutoff for per sample m6A score; if >0, then do filtering.}
}
\value{
A filtered data frame.
}
\description{
This function calculates the CV of columns m6A and RUD, and filters highly variable genes.
}
\examples{
\dontrun{
mp1=filterHVPerGene(mp, m6Acutoff=0.3, RUDcutoff=0.3,
                    totalWeights=100, perWeights=10, nWeights=6)
}
}
