% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{FitModel}
\alias{FitModel}
\title{FitModel}
\usage{
FitModel(
  type = "fitError",
  gene = NA_character_,
  df.residual = NA_real_,
  dispersion = NA_real_,
  coefficients = NA_real_,
  raw.pvals = NA_real_,
  cov.unscaled = matrix(NA, 0, 0),
  var.post = NA_real_,
  df.post = NA_real_
)
}
\arguments{
\item{type}{default set to fitError, can be a glm}

\item{gene}{gene}

\item{df.residual}{Numeric, residual of the fitted glm}

\item{dispersion}{Numeric, dispersion of the fitted glm}

\item{coefficients}{Numeric, coefficients of the fitted glm}

\item{raw.pvals}{Numeric, p-values of the fitted glm}

\item{cov.unscaled}{Numeric, unscaled variance of the fitted glm}

\item{var.post}{Numeric, posterior variance of the glm, default is NA}

\item{df.post}{Numeric, posterior degrees of freedom of the glm, default is NA}
}
\value{
A FitModel object
}
\description{
Function for contstructing a new `FitModel` object.
}
\seealso{
Other QBGLM functions:
\code{\link[=QBGLM]{QBGLM()}},
\code{\link[=QBGLMmultiDs]{QBGLMmultiDs()}},
\code{\link[=addPost2FitModels]{addPost2FitModels()}},
\code{\link[=fitQuasiGLM]{fitQuasiGLM()}},
\code{\link[=getM6AperGene]{getM6AperGene()}},
\code{\link[=getRUDperGene]{getRUDperGene()}},
\code{\link[=removeFitErrorModels]{removeFitErrorModels()}},
\code{\link[=testBetas]{testBetas()}}
}
\concept{QBGLM functions}
