% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{fitQuasiGLM}
\alias{fitQuasiGLM}
\title{Fit QBGLM for genes}
\usage{
fitQuasiGLM(mpMat, noFitError = TRUE)
}
\arguments{
\item{mpMat}{A data frame of genes with m6A+RUD+Weights, there should be columns ending with "_m6A", "_RUD", "_weight".}

\item{noFitError}{TRUE to remove fitError models.}
}
\value{
A list of fitted models with each element a fitted GLM for a gene.
}
\description{
This function fits QBGLM for a data frame of genes with m6A+RUD+Weights, and returns a model list with each element a fitted GLM for a gene.
}
\examples{
\dontrun{
p=getRUDperGene(apadsPD)
m=getM6AperGene(m6ads, apadsPD)
mp=merge(m, p, by.x='gene', by.y='gene')
models=fitQuasiGLM(mp)
table(unlist(lapply(models, function(par) par@type)))
}
}
\seealso{
Other QBGLM functions:
\code{\link[=FitModel]{FitModel()}},
\code{\link[=QBGLM]{QBGLM()}},
\code{\link[=QBGLMmultiDs]{QBGLMmultiDs()}},
\code{\link[=addPost2FitModels]{addPost2FitModels()}},
\code{\link[=getM6AperGene]{getM6AperGene()}},
\code{\link[=getRUDperGene]{getRUDperGene()}},
\code{\link[=removeFitErrorModels]{removeFitErrorModels()}},
\code{\link[=testBetas]{testBetas()}}
}
\concept{QBGLM functions}
