% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{get3UTRAPApd}
\alias{get3UTRAPApd}
\title{Get proximal and distal polyA sites in a PACdataset}
\usage{
get3UTRAPApd(
  pacds,
  minDist = 50,
  maxDist = 1000,
  minRatio = 0.05,
  fixDistal = FALSE,
  addCols = "pd"
)
}
\arguments{
\item{pacds}{A PACdataset object. By default, all are 3'UTR polyA sites; If there is an `ftr` column in PACds, then will check if all ftr are 3'UTR.}

\item{minDist}{Default is 50. Require the distance between  proximal and distal polyA site to be > minDist and < maxDist.}

\item{maxDist}{Default is 1000. Require the distance between  proximal and distal polyA site to be > minDist and < maxDist.}

\item{minRatio}{Default is 0.05. If >0 then will filter out polyA sites that do not meet the requirements based on the ratio first, and then filter proximal and distal sites.}

\item{fixDistal}{Default is FALSE. True to fix the farthest PA, that is, only look for the proximal site.}

\item{addCols}{Default is `pd`. If not NULL, then will add columns naming <addCols>Which/Score/Ratio/Dist, like pdWhich=P or D.}
}
\value{
A PACdataset with derived gram and distance columns added.
}
\description{
get3UTRAPApd gets proximal and distal polyA sites in a PACdataset
Selection criteria: among all PA pairs that meet the criteria of `dist` and `ratio`, select the pair with the highest ratio sum and the closest distance.
Note that there is no filtering of PA expression levels here, which can be done in movAPA::subsetPACds.
}
\examples{
\dontrun{
pacds=APAtrap2PACds(paCtrl)
pd1=get3UTRAPApd(pacds, minDist=50, maxDist=1000, minRatio=0.05, fixDistal=F)
pd2=get3UTRAPApd(pacds, minDist=50, maxDist=1000, minRatio=0.05, fixDistal=T)
pacds@anno$gene[!(pacds@anno$gene \%in\% pd1@anno$gene)]
movAPA::subsetPACds(pd1, gene='CTC1')
movAPA::subsetPACds(pd2, gene='CTC1')
}
}
\seealso{
Other APAds functions:
\code{\link[=APAds2GLMds]{APAds2GLMds()}},
\code{\link[=QAPA2GLMds]{QAPA2GLMds()}},
\code{\link[=addM6A2PACds]{addM6A2PACds()}},
\code{\link[=addMMPAS2PACds]{addMMPAS2PACds()}},
\code{\link[=addPAS2PACds]{addPAS2PACds()}},
\code{\link[=get3UTRAPApdDE]{get3UTRAPApdDE()}},
\code{\link[=mergePACdsCoords]{mergePACdsCoords()}},
\code{\link[=mergePACdsRanges]{mergePACdsRanges()}}
}
\concept{APAds functions}
