% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{get3UTRAPApdDE}
\alias{get3UTRAPApdDE}
\title{Get 3'UTR DEAPA from a 3UTRAPApd-like PACdataset.}
\usage{
get3UTRAPApdDE(
  pacds,
  pthd = 0.05,
  scol = "group",
  DEcol = "isDE",
  pvcol = "DEPval",
  filterDE = TRUE
)
}
\arguments{
\item{pacds}{A PACdataset object with `pdWhich` column which is obtained by `get3UTRAPApd`.}

\item{pthd}{P-value cutoff of Fisher.test to filter DEAPA, default is 0.05.}

\item{scol}{The column denoting "group" in pacds@colData, default is 'group'.}

\item{DEcol}{Default is 'isDE', Will add isDE=TRUE/FALSE to pacds@anno.}

\item{pvcol}{Default is 'DEPval', Will add DEPval to pacds@anno.}

\item{filterDE}{Default is TRUE to filter isDE=TRUE rows.}
}
\value{
A PACdataset with DE information or filtered.
}
\description{
get3UTRAPApdDE gets 3'UTR DEAPA from a 3UTRAPApd-like PACdataset. This function filters DEAPA by fisher.test for each replicates of each pair of conditions.
As long as one replicate of a pair of conditions satisfies pval<cutoff, then it is marked as DEAPA.
This function determines the number and name of conditions, replicates based on `scol`, and consider rep1 and rep2 for each pair of condition pairs in order.
}
\examples{
\dontrun{
apadsUTR_m3=movAPA::get3UTRAPAds(apads_m3)
apadsPD_m3=get3UTRAPApd(apadsUTR_m3, minDist=50, maxDist=5000, minRatio=0.05,
                       fixDistal=FALSE, addCols='pd')
apadsUTR_m3DE=get3UTRAPApdDE(apadsPD_m3, pthd=0.05, filterDE=TRUE)
}
}
\seealso{
Other APAds functions:
\code{\link[=APAds2GLMds]{APAds2GLMds()}},
\code{\link[=QAPA2GLMds]{QAPA2GLMds()}},
\code{\link[=addM6A2PACds]{addM6A2PACds()}},
\code{\link[=addMMPAS2PACds]{addMMPAS2PACds()}},
\code{\link[=addPAS2PACds]{addPAS2PACds()}},
\code{\link[=get3UTRAPApd]{get3UTRAPApd()}},
\code{\link[=mergePACdsCoords]{mergePACdsCoords()}},
\code{\link[=mergePACdsRanges]{mergePACdsRanges()}}
}
\concept{APAds functions}
