% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{getFitDataByGenes}
\alias{getFitDataByGenes}
\title{Get genes' GLM-fitted results}
\usage{
getFitDataByGenes(fitRes, genes)
}
\arguments{
\item{fitRes}{A data frame from `testBetas`, with rawPval1, pval1 and (qval1, padj1...).}

\item{genes}{genes to be subset.}
}
\value{
A data frame.
}
\description{
getFitDataByGenes gets the m6A/RUD/weight data frame of given gene names  (which can be the gene, gene assembly, gene symbol column).
}
\examples{
\dontrun{
getFitDataByGenes(res, genes=c(7114, 2107))
}
}
\seealso{
Other output functions:
\code{\link[=statFitRes]{statFitRes()}},
\code{\link[=subsetFitRes]{subsetFitRes()}}
}
\concept{output functions}
