% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{getGeneAnnos}
\alias{getGeneAnnos}
\title{Get genomic regions of all genes}
\usage{
getGeneAnnos(annoObj)
}
\arguments{
\item{annoObj}{An OrganismDb or Mart object.}
}
\value{
A data frame of genomic ranges of all genes, with fixed columns of chr/strand/start/end and other columns starting with 'gene_' like gene_entrez.
}
\description{
getGeneAnnos retrieves genomic ranges of all genes from different genome annotation sources.
}
\examples{
\dontrun{
library(Homo.sapiens)
orgdb=Homo.sapiens
geneAnnos=getGeneAnnos(orgdb)
}
}
