% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{getM6AperGene}
\alias{getM6AperGene}
\title{Calculate gene level m6A score}
\usage{
getM6AperGene(m6Ads, pacds, sufix = "_m6A")
}
\arguments{
\item{m6Ads}{A PACdataset object of m6A sites.}

\item{pacds}{A PACdataset object of polyA sites.
Both m6ads and pacds are required to have a gene column.
We can first annotate both `ds` with 3' UTR extension (ensuring m6A and polyA sites are all within the same gene), and then call this function.}

\item{sufix}{Suffix for the columns names of m6A score, default is '_m6A'.}
}
\value{
A data frame with columns like [gene, d0, nM6A, ...samples in m6Ads@counts with sufix _m6A].
}
\description{
This function first searches the closest polyA site within the same gene for each m6A site,
and calculate the m6A score by distance between polyA site and m6A using M=exp (- d/d0).
d0 is the quantile of 75% of all nearest m6A and polyA site distance.
Then the gene level m6A score is the average m6A level of all m6As in a gene after distance weighting.
}
\examples{
m6Ads=movAPA::makeExamplePACds()
m6Ads@anno$gene=sample(LETTERS, length(m6Ads), replace=TRUE)
pacds=movAPA::makeExamplePACds()
pacds@anno$gene=sample(LETTERS, length(pacds), replace=TRUE)
m=getM6AperGene(m6Ads, pacds)
head(m)
}
\seealso{
Other QBGLM functions:
\code{\link[=FitModel]{FitModel()}},
\code{\link[=QBGLM]{QBGLM()}},
\code{\link[=QBGLMmultiDs]{QBGLMmultiDs()}},
\code{\link[=addPost2FitModels]{addPost2FitModels()}},
\code{\link[=fitQuasiGLM]{fitQuasiGLM()}},
\code{\link[=getRUDperGene]{getRUDperGene()}},
\code{\link[=removeFitErrorModels]{removeFitErrorModels()}},
\code{\link[=testBetas]{testBetas()}}
}
\concept{QBGLM functions}
