% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{getRUDperGene}
\alias{getRUDperGene}
\title{Calculate gene level RUD score}
\usage{
getRUDperGene(pacdsPD, sufix1 = "_RUD", sufix2 = "_weight")
}
\arguments{
\item{pacdsPD}{A PACdataset object with only proximal and distal polyA sites, which could be obtained by `get3UTRAPApd`.}

\item{sufix1}{Suffix for the columns names of RUD score, default is '_RUD'.}

\item{sufix2}{Suffix for the columns names of counts, default is '_weight'.}
}
\value{
A data frame with columns like [pacdsPD@counts with sufix _RUD and _weight].
}
\description{
This function calculates gene level RUD score.
}
\examples{
\dontrun{
apadsUTR=movAPA::get3UTRAPAds(apads)
apadsPD=get3UTRAPApd(apadsUTR, minDist=50, maxDist=5000, minRatio=0.05,
                    fixDistal=FALSE, addCols='pd')
p=getRUDperGene(apadsPD)
}
}
\seealso{
Other QBGLM functions:
\code{\link[=FitModel]{FitModel()}},
\code{\link[=QBGLM]{QBGLM()}},
\code{\link[=QBGLMmultiDs]{QBGLMmultiDs()}},
\code{\link[=addPost2FitModels]{addPost2FitModels()}},
\code{\link[=fitQuasiGLM]{fitQuasiGLM()}},
\code{\link[=getM6AperGene]{getM6AperGene()}},
\code{\link[=removeFitErrorModels]{removeFitErrorModels()}},
\code{\link[=testBetas]{testBetas()}}
}
\concept{QBGLM functions}
