% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{getVarGrams}
\alias{getVarGrams}
\title{Get hexamers of polyA signals.}
\usage{
getVarGrams(grams)
}
\arguments{
\item{grams}{the grams vector or be V1/MM/MOUSE/MM10.}
}
\value{
grams of uppercase.
}
\description{
Get hexamers of polyA signals.
}
\examples{
getVarGrams('v1')
getVarGrams('MM')
getVarGrams(c('aataaa','ggattc'))
}
