% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{m6A2PACds}
\alias{m6A2PACds}
\title{Convert a m6A data frame to PACdataset.}
\usage{
m6A2PACds(d)
}
\arguments{
\item{d}{A data frame with columns: seqnames, geneStrand, start, end, width, geneId, score, geneStart, geneEnd.}
}
\value{
A PACdataset object. The coord is peak_start(+) or peak_end(-), add will add UPA_start, UPA_end columns.
}
\description{
m6A2PACds convert a m6A data frame to PACdataset.
}
\seealso{
Other m6Adata functions:
\code{\link[=m6ADB2PACds]{m6ADB2PACds()}},
\code{\link[=m6AExpressDf2PACds]{m6AExpressDf2PACds()}},
\code{\link[=m6Ads2GLMds]{m6Ads2GLMds()}},
\code{\link[=m6Apeak2GLMds]{m6Apeak2GLMds()}}
}
\concept{m6Adata functions}
