% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{m6ADB2PACds}
\alias{m6ADB2PACds}
\title{convert m6ADB data to PACdataset}
\usage{
m6ADB2PACds(d)
}
\arguments{
\item{d}{A file name of m6ADB data.
Columns starting with `RPM.` is the sample columns.
coord=peak_start(+)/peak_end(-), add will add peak_score, UPA_start and UPA_end.}
}
\value{
PACdataset.
}
\description{
m6ADB2PACds convert m6ADB data to PACdataset.
}
\examples{
\dontrun{
d=m6ADB2PACds('human_GSE46705_HeLa_M3KO_Ctrl.csv')
}
}
\seealso{
Other m6Adata functions:
\code{\link[=m6A2PACds]{m6A2PACds()}},
\code{\link[=m6AExpressDf2PACds]{m6AExpressDf2PACds()}},
\code{\link[=m6Ads2GLMds]{m6Ads2GLMds()}},
\code{\link[=m6Apeak2GLMds]{m6Apeak2GLMds()}}
}
\concept{m6Adata functions}
