% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{m6Ads2GLMds}
\alias{m6Ads2GLMds}
\title{Wrapper function to load m6A's PACdataset to PACdataset for QBGLM.}
\usage{
m6Ads2GLMds(m6Ads, DEpval = 0.05, txdb, extUTRlen = 2000)
}
\arguments{
\item{m6Ads}{A pacdataset of m6A peaks.}

\item{DEpval}{cutoff to filter DE-m6A, default is 0.05. If =1, then not filter DE.}

\item{txdb}{An TXDB object for genome annotation.}

\item{extUTRlen}{Length to extend 3'UTR to include more near downstream m6A sites.}
}
\value{
A PACdataset of m6A peaks or DE-m6A peaks (if DEpval<1), with @anno have raw_<IP, Input count columns>, DE_log.OR/pvalue/padj.
}
\description{
m6Ads2GLMds is a wrapper function to load m6A's PACdataset to PACdataset for QBGLM. It annotates m6A sites with TXDB. Finally, it filters DE-m6As.
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
txdb=TxDb.Hsapiens.UCSC.hg38.knownGene
m6Afile='M3KD.m6A.txt'

# libsize of m6Aexpress, with two columns: sample and libsize.
libSizes=read.table('M3KD.m6A.libsize.txt', header=TRUE)

# new sample and group names for samples in libSizes
newSampleNames=c('ctrlM31','ctrlM32','treatM31','treatM32')
newSampleGroups=c('ctrlM3','ctrlM3','treatM3','treatM3')

# get m6A for QBGLM
m6ads=m6A2GLMds(m6Afile, libSizes,
                newSampleNames, newSampleGroups,
                DEpval=0.05,
                txdb=txdb,
                extUTRlen=2000)

movAPA::summary(m6ads)
m6Ads2GLMds(m6Ads=m6ads, DEpval=0.05, txdb, extUTRlen=2000)
}
}
\seealso{
Other m6Adata functions:
\code{\link[=m6A2PACds]{m6A2PACds()}},
\code{\link[=m6ADB2PACds]{m6ADB2PACds()}},
\code{\link[=m6AExpressDf2PACds]{m6AExpressDf2PACds()}},
\code{\link[=m6Apeak2GLMds]{m6Apeak2GLMds()}}
}
\concept{m6Adata functions}
