% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{m6AExpressDf2PACds}
\alias{m6AExpressDf2PACds}
\title{Convert m6AExpress's data frame to PACdataset.}
\usage{
m6AExpressDf2PACds(d, smpCols = NULL)
}
\arguments{
\item{d}{A m6AExpress's data frame, with columns: seqnames, start, end, width, strand, gene_name, and <IP1, IP2, Treated_IP1, Treated_IP2..>.}

\item{smpCols}{If not NULL then will extract `smpCols`, otherwise will auto detect sample columns as remaining columns.}
}
\value{
A PACdataset object. The coord is peak_start(+) or peak_end(-), add will add UPA_start, UPA_end columns.
}
\description{
m6AExpressDf2PACds converts m6AExpress's data frame to PACdataset.
}
\examples{
\dontrun{
m6ads=m6AExpressDf2PACds(peakDfNorm,
                   smpCols=c("IP1", "IP2", "Treated_IP1", "Treated_IP2"))
}
}
\seealso{
Other m6Adata functions:
\code{\link[=m6A2PACds]{m6A2PACds()}},
\code{\link[=m6ADB2PACds]{m6ADB2PACds()}},
\code{\link[=m6Ads2GLMds]{m6Ads2GLMds()}},
\code{\link[=m6Apeak2GLMds]{m6Apeak2GLMds()}}
}
\concept{m6Adata functions}
