% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{m6AIpInputDE}
\alias{m6AIpInputDE}
\title{Use QNBtest to test DE m6A peak}
\usage{
m6AIpInputDE(peakDf, ctrls, treats, libSizes = NULL, prefix = "DE_")
}
\arguments{
\item{peakDf}{A data frame of peaks.}

\item{ctrls}{cond1's IP and Input column names or idx.}

\item{treats}{cond2's IP and Input column names or idx.}

\item{libSizes}{Total read counts for each ctrls/treats (same order as ctrls/treats); if NULL then use the colSums.}

\item{prefix}{Default is `DE_`, which will add <DE_> to columns names like DE_log2.OR, DE_pvalue, DE_padj.}
}
\value{
The given peak data frame with three new cols log2.OR, pvalue, padj.
}
\description{
m6AIpInputDE  use QNBtest to test DE m6A peak
This function searches the nearest m6A upstream of each polyA site and record the distance, coordinate of m6A.
This function will add to pacds@anno the m6A_ovp/dist.. columns. Dist is from peak_start of m6A to coord of the polyA site.
}
\examples{
\dontrun{
dat=.getRdaData('m3mut_cond1.rda')
m6adsM3=m6AExpress2PACds(peakDf=dat[[1]], ctrls=dat[[2]],
                        treats=dat[[3]], libSizes=dat[[4]],
                       doDE=TRUE, filterDE=FALSE)
peakDf=m6AIpInputDE(peakDf, ctrls=ctrls, treats=treats,
                    libSizes=libSizes, prefix='DE_')
}
}
\seealso{
Other m6Ads functions:
\code{\link[=m6AIpInputNormalize]{m6AIpInputNormalize()}}
}
\concept{m6Ads functions}
