% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{m6AIpInputNormalize}
\alias{m6AIpInputNormalize}
\title{Normalize m6A levels}
\usage{
m6AIpInputNormalize(peakDf, ctrls, treats, libSizes = NULL, rawPrefix = "raw_")
}
\arguments{
\item{peakDf}{A data frame of peaks.}

\item{ctrls}{cond1's IP and Input column names or idx.}

\item{treats}{cond2's IP and Input column names or idx.}

\item{libSizes}{Total read counts for each ctrls/treats (same order as ctrls/treats); if NULL then use the colSums.}

\item{rawPrefix}{Default is `raw_`. If not NULL to keep raw counts of IP and Input, but adding 'raw_'.}
}
\value{
The given peak data frame with raw_ctrls/treats + (IP column names denoting normalized columns).
}
\description{
m6AIpInputNormalize gets normalized m6A levels.
After standardizing the IP and Input columns according to libsize, the IP/Input of each rep is then taken as log2.
In m6Aexpress, M6A level=log2 (Cip/Sip) - log2 (Cinput/Input), where Sip and Input are normalization factors; Cip is the m6A level, and input is the background level.
}
\examples{
\dontrun{
peakDfNorm=m6AIpInputNormalize(peakDf, ctrls=ctrls, treats=treats,
                              libSizes=libSizes, rawPrefix='raw_')
}
}
\seealso{
Other m6Ads functions:
\code{\link[=m6AIpInputDE]{m6AIpInputDE()}}
}
\concept{m6Ads functions}
