% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{m6Apeak2GLMds}
\alias{m6Apeak2GLMds}
\title{Wrapper function to load m6Aexpress' results to PACdataset for QBGLM.}
\usage{
m6Apeak2GLMds(
  m6Afile,
  libSizes,
  newSampleNames,
  newSampleGroups,
  DEpval = 0.05,
  txdb,
  extUTRlen = 2000
)
}
\arguments{
\item{m6Afile}{A m6AExpress's m6A peak file, with columns seqnames, start, end, width, strand, gene_name, and <IP1, IP2, Treated_IP1, Treated_IP2..>.}

\item{libSizes}{A file name or a data frame with 2 columns: sample and counts. The first column is the sample column, with the order -- ctrls: IP, Input; treats: IP, Input.
The second columns is the total read counts for each sample.}

\item{newSampleNames}{new sample names, e.g., c('ctrlM31','ctrlM32','treatM31','treatM32'), should be as the order of libSizes.}

\item{newSampleGroups}{new group name, e.g., c('ctrlM3','ctrlM3','treatM3','treatM3'), should be as the order of libSizes.}

\item{DEpval}{cutoff to filter DE-m6A, default is 0.05. If =1, then not filter DE.}

\item{txdb}{An TXDB object for genome annotation.}

\item{extUTRlen}{Length to extend 3'UTR to include more near downstream m6A sites.}
}
\value{
A PACdataset of m6A peaks or DE-m6A peaks (if DEpval<1), with @anno have raw_<IP, Input count columns>, DE_log.OR/pvalue/padj.
}
\description{
m6Apeak2GLMds is a wrapper function to load m6Aexpress' results to PACdataset for QBGLM. It first get m6A scores by reading m6Afile and libSizes.
Then it annotates m6A sites with TXDB. Finally, it filters DE-m6As.
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
txdb=TxDb.Hsapiens.UCSC.hg38.knownGene
m6Afile='M3KD.m6A.txt'
libSizes=read.table('M3KD.m6A.libsize.txt', header=TRUE)
newSampleNames=c('ctrlM31','ctrlM32','treatM31','treatM32')
newSampleGroups=c('ctrlM3','ctrlM3','treatM3','treatM3')
m6A2GLMds(m6Afile, libSizes,
          newSampleNames, newSampleGroups,
          DEpval=0.05,
          txdb=txdb,
          extUTRlen=2000)
}
}
\seealso{
Other m6Adata functions:
\code{\link[=m6A2PACds]{m6A2PACds()}},
\code{\link[=m6ADB2PACds]{m6ADB2PACds()}},
\code{\link[=m6AExpressDf2PACds]{m6AExpressDf2PACds()}},
\code{\link[=m6Ads2GLMds]{m6Ads2GLMds()}}
}
\concept{m6Adata functions}
