% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{mergePACdsCoords}
\alias{mergePACdsCoords}
\title{Merge multiple PACdataset objects by cleavage-site coordinates}
\usage{
mergePACdsCoords(PACdsList, d = 24)
}
\arguments{
\item{PACdsList}{A `PACdataset` object or a list of `PACdataset` objects.
Each `PACds@anno` must contain `chr`, `strand`, and `coord` columns. If
`colData` is absent, sample groups are assigned as `group1`, `group2`, and
so on.}

\item{d}{Distance used to group nearby PACs. Default is 24 nt.}
}
\value{
A merged `PACdataset` object. The `counts` slot stores counts of
  merged samples. The `colData` slot stores group annotations derived from
  the first column of each input object's `colData`. The `anno` slot includes
  `chr`, `strand`, `coord`, `tottag`, `UPA_start`, `UPA_end`, `nPA`, and
  `maxtag`.
}
\description{
mergePACdsCoords groups nearby poly(A) sites from one or multiple
`PACdataset` objects. This function was modified from
`movAPA::mergePACds` and aggregates pAs according to their `coord` values.
}
\details{
This function is useful for grouping nearby cleavage sites into PACs. It can
also merge multiple PA or PAC datasets into a single `PACdataset` for
differential APA or other downstream analyses. After grouping or merging,
the returned object can be annotated using `annotatePAC()`.
}
\examples{
\dontrun{
# Group nearby pAs into PACs.
data(PACds)
PACds@counts <- rbind(PACds@counts, PACds@counts)
PACds@anno <- rbind(PACds@anno, PACds@anno)
ds <- mergePACdsCoords(PACds, d = 24)

# Merge two PACdataset objects.
pacds <- new("PACdataset", counts = PACds@counts, anno = PACds@anno)
PACdsList <- list(PACds, pacds)
ds <- mergePACdsCoords(PACdsList, d = 24)
}

}
\seealso{
Other APAds functions:
\code{\link[=APAds2GLMds]{APAds2GLMds()}},
\code{\link[=QAPA2GLMds]{QAPA2GLMds()}},
\code{\link[=addM6A2PACds]{addM6A2PACds()}},
\code{\link[=addMMPAS2PACds]{addMMPAS2PACds()}},
\code{\link[=addPAS2PACds]{addPAS2PACds()}},
\code{\link[=get3UTRAPApd]{get3UTRAPApd()}},
\code{\link[=get3UTRAPApdDE]{get3UTRAPApdDE()}},
\code{\link[=mergePACdsRanges]{mergePACdsRanges()}}
}
\concept{APAds functions}
