% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{mergePACdsRanges}
\alias{mergePACdsRanges}
\title{Merge multiple PACdataset}
\usage{
mergePACdsRanges(PACdsList, d = 0)
}
\arguments{
\item{PACdsList}{A `PACdataset` object or a list of multiple `PACdataset`
objects. `PACds@anno` must contain the `chr`, `strand`, `UPA_start`, and
`UPA_end` columns. If names are provided, the corresponding PA IDs are
recorded using these names; otherwise, names are assigned as `ds1`, `ds2`,
and so on.}

\item{d}{Allowed gap between peaks in PACds. If `d = 0`, two peaks must
overlap by at least 1 nt to be merged. Default is 24 nt.}
}
\value{
A merged `PACdataset` object with annotation columns including
  `chr`, `strand`, `tottag`, `UPA_start`, `UPA_end`, `nPA`, input-dataset
  PA IDs, and `coord`.
}
\description{
mergePACdsRanges groups nearby PACs from single/multiple PACdataset objects.
This function was modified from movAPA::mergePACds, aggregating pAs according
to genomic ranges. Peaks are merged according to `UPA_start` and `UPA_end`,
where `d` is the allowed gap. The PA ID from each input PACdataset is recorded
for each merged peak.
}
\examples{
\dontrun{
# Merge two PACdataset objects
pacds <- new("PACdataset", counts = PACds@counts, anno = PACds@anno)
PACdsList <- list(PACds, pacds)
ds <- mergePACdsRanges(PACdsList, d = 24)
}

}
\seealso{
Other APAds functions:
\code{\link[=APAds2GLMds]{APAds2GLMds()}},
\code{\link[=QAPA2GLMds]{QAPA2GLMds()}},
\code{\link[=addM6A2PACds]{addM6A2PACds()}},
\code{\link[=addMMPAS2PACds]{addMMPAS2PACds()}},
\code{\link[=addPAS2PACds]{addPAS2PACds()}},
\code{\link[=get3UTRAPApd]{get3UTRAPApd()}},
\code{\link[=get3UTRAPApdDE]{get3UTRAPApdDE()}},
\code{\link[=mergePACdsCoords]{mergePACdsCoords()}}
}
\concept{APAds functions}
