% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{QAPA2GLMds}
\alias{QAPA2GLMds}
\title{Wrapper function to load QAPA results to PACdataset for QBGLM.}
\usage{
QAPA2GLMds(
  qapaFile,
  sfFiles,
  newSampleNames,
  newSampleGroups,
  PDminDist = 50,
  PDmaxDist = 5000,
  PDminRatio = 0.05,
  PDfixDistal = FALSE,
  DEpval = 0.05,
  txdb,
  extUTRlen = 2000
)
}
\arguments{
\item{qapaFile}{A file name of the QAPA APA list.}

\item{sfFiles}{A vector of quant.sf files of QAPA's outputs which record raw counts. Names of `sfFiles` should be sample names in the `qapaFile`.}

\item{newSampleNames}{new sample names for the sfFiles, e.g., c('ctrlM31','ctrlM32','treatM31','treatM32').}

\item{newSampleGroups}{new group names for the sfFiles, e.g., c('ctrlM3','ctrlM3','treatM3','treatM3').}

\item{PDminDist}{Default is 50. Require the distance between  proximal and distal polyA site to be > minDist and < maxDist.}

\item{PDmaxDist}{Default is 1000. Require the distance between  proximal and distal polyA site to be > minDist and < maxDist.}

\item{PDminRatio}{Default is 0.05. If >0 then will filter out polyA sites that do not meet the requirements based on the ratio first, and then filter proximal and distal sites.}

\item{PDfixDistal}{Default is FALSE. True to fix the farthest PA, that is, only look for the proximal site.}

\item{DEpval}{cutoff to filter DE-APA, default is 0.05. If =1, then not filter DE.}

\item{txdb}{An TXDB object for genome annotation.}

\item{extUTRlen}{Length to extend 3'UTR to include more near downstream polyA sites.}
}
\value{
A PACdataset with only proximal and distal, and DE-APA (if DEpval<1) sites.
}
\description{
QAPA2GLMds is a wrapper function to load QAPA results to PACdataset for QBGLM. It first get raw counts of polyA sites by reading qapafile and quant.sf files.
Then it annotates polyA sites with TXDB and get proximal and distal sites of 3'UTR-APA genes. Finally, it filters DE-APA by fisher.test of replicates.
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
txdb=TxDb.Hsapiens.UCSC.hg38.knownGene
qapaFile='M3KD.qapa.txt'
sfFiles=c('qapa_SRR847370_quant.sf', 'qapa_SRR847371_quant.sf',
         'qapa_SRR847374_quant.sf', 'qapa_SRR847375_quant.sf')
names(sfFiles)=paste0('sample', 1:4)
newSampleNames=c('ctrlM31','ctrlM32','treatM31','treatM32')
newSampleGroups=c('ctrlM3','ctrlM3','treatM3','treatM3')
apads=QAPA2GLMds(qapaFile, sfFiles,
          newSampleNames, newSampleGroups,
          PDminDist=50, PDmaxDist=5000, PDminRatio=0.05, PDfixDistal=FALSE,
          DEpval=0.05,
          txdb=txdb,
          extUTRlen=2000)
}
}
\seealso{
Other APAds functions:
\code{\link[=APAds2GLMds]{APAds2GLMds()}},
\code{\link[=addM6A2PACds]{addM6A2PACds()}},
\code{\link[=addMMPAS2PACds]{addMMPAS2PACds()}},
\code{\link[=addPAS2PACds]{addPAS2PACds()}},
\code{\link[=get3UTRAPApd]{get3UTRAPApd()}},
\code{\link[=get3UTRAPApdDE]{get3UTRAPApdDE()}},
\code{\link[=mergePACdsCoords]{mergePACdsCoords()}},
\code{\link[=mergePACdsRanges]{mergePACdsRanges()}}
}
\concept{APAds functions}
