% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{QAPA2PACds}
\alias{QAPA2PACds}
\title{convert QAPA results to PACdataset}
\usage{
QAPA2PACds(d, vcol = "TPM")
}
\arguments{
\item{d}{A file name of QAPA output or a data.frame object.
The file or data.frame contains columns: Transcript Gene Gene_Name Chr Strand UTR3.Start UTR3.End, and sample columns like "sample1.PAU sample2.PAU..." or "sample1.TPM sample2.TPM ...".}

\item{vcol}{The string to extract sample columns. For example, vcol='TPM' means to get all columns containing 'TPM'.}
}
\value{
PACdataset with all-zero rows removed.
}
\description{
QAPA2PACds convert QAPA results to PACdataset.
}
\examples{
\dontrun{
apads=QAPA2PACds('sim3_qapa_hg38.txt', vcol='TPM')
}
}
\seealso{
Other APAdata functions:
\code{\link[=APAtrap2PACds]{APAtrap2PACds()}},
\code{\link[=addQAPARawCounts]{addQAPARawCounts()}},
\code{\link[=bed2PACds]{bed2PACds()}}
}
\concept{APAdata functions}
