% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{QBGLM}
\alias{QBGLM}
\title{Wrapper function of QBGLM}
\usage{
QBGLM(APAds, m6Ads)
}
\arguments{
\item{APAds}{A PACdataset from QAPA2GLMds, txdb-annotated DE-PD-APA dataset.}

\item{m6Ads}{A PACdataset from m6A2GLMds, txdb-annotated DE-m6A dataset.}
}
\value{
A data frame with each row a gene.
}
\description{
QBGLM is a wrapper function of QBGLM using APA from `QAPA2GLMds` and m6A `m6A2GLMds`. It first links APAds and m6Ads to get per gene data.
And then do QBGLM to test each gene and get betas.
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
txdb=TxDb.Hsapiens.UCSC.hg38.knownGene

# output APA list from QAPA, with TPM for each sample
qapaFile='M3KD.qapa.txt'

# quant.sf: raw counts of salmon from QAPA's output
sfFiles=c('qapa_SRR847370_quant.sf', 'qapa_SRR847371_quant.sf',
          'qapa_SRR847374_quant.sf', 'qapa_SRR847375_quant.sf')
names(sfFiles)=paste0('sample', 1:4)

# new sample and group names for sfFiles
newSampleNames=c('ctrlM31','ctrlM32','treatM31','treatM32')
newSampleGroups=c('ctrlM3','ctrlM3','treatM3','treatM3')

# get APA for QBGLM
apads=QAPA2GLMds(qapaFile=qapaFile, sfFiles=sfFiles,
                 newSampleNames=newSampleNames, newSampleGroups=newSampleGroups,
                 PDminDist=50, PDmaxDist=5000, PDminRatio=0.05, PDfixDistal=FALSE,
                 DEpval=0.05,
                 txdb=txdb,
                 extUTRlen=2000)
movAPA::summary(apads)


# A table of Get_peak_infor of m6Aexpress.
m6Afile='M3KD.m6A.txt'

# libsize of m6Aexpress, with two columns: sample and libsize.
libSizes=read.table('M3KD.m6A.libsize.txt', header=TRUE)

# new sample and group names for samples in libSizes
newSampleNames=c('ctrlM31','ctrlM32','treatM31','treatM32')
newSampleGroups=c('ctrlM3','ctrlM3','treatM3','treatM3')

# get m6A for QBGLM
m6ads=m6A2GLMds(m6Afile, libSizes,
                newSampleNames, newSampleGroups,
                DEpval=0.05,
                txdb=txdb,
                extUTRlen=2000)

movAPA::summary(m6ads)

# QBGLM
resMp=QBGLM(APAds=apads, m6Ads=m6ads)
nrow(resMp)
# just simplify columns without filtering
resMp=subsetFitRes(resMp, qval1=2)

## Add gene symbols and entrezids
library(Homo.sapiens)
orgdb=Homo.sapiens
geneAnnos=getAnnoGenes(orgdb)
resMp=merge(resMp,
           geneAnnos[, c('gene_entrezid','gene_ensembl','gene_symbol')],
           all.x=TRUE, all.y=FALSE, by.x='gene', by.y='gene_entrezid')
head(resMp)
write.csv(resMp, file='M3KD_HeLa1.resMp.csv')

## subset genes with qval<0.2
resMpSig=subsetFitRes(resMp, qval1=0.2)
write.csv(resMpSig, file='M3KD_HeLa1.m6APAreg.csv')
}
}
\seealso{
Other QBGLM functions:
\code{\link[=FitModel]{FitModel()}},
\code{\link[=QBGLMmultiDs]{QBGLMmultiDs()}},
\code{\link[=addPost2FitModels]{addPost2FitModels()}},
\code{\link[=fitQuasiGLM]{fitQuasiGLM()}},
\code{\link[=getM6AperGene]{getM6AperGene()}},
\code{\link[=getRUDperGene]{getRUDperGene()}},
\code{\link[=removeFitErrorModels]{removeFitErrorModels()}},
\code{\link[=testBetas]{testBetas()}}
}
\concept{QBGLM functions}
