% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{QBGLMmultiDs}
\alias{QBGLMmultiDs}
\title{Wrapper function of QBGLM for multiple samples}
\usage{
QBGLMmultiDs(
  APAdsList,
  m6AdsList,
  distAPA = 24,
  distM6A = 50,
  txdb,
  extUTRlen = 2000,
  DEm6Apval = 0.05,
  PDminDist = 50,
  PDmaxDist = 5000,
  PDminRatio = 0.05,
  PDfixDistal = FALSE,
  DEAPApval = 0.05
)
}
\arguments{
\item{APAdsList}{A PACdataset list of APAds from functions like `QAPA2PACds`, without annotation and DE calculation.}

\item{m6AdsList}{A PACdataset list of m6Ads from `m6AExpress2PACds`, which have done DE.}

\item{distAPA}{distance to group nearby polyA sites in APAds, default is 24 nt.}

\item{distM6A}{distance to group nearby m6A sites in m6Ads, default is 50 nt.}

\item{txdb}{An TXDB object for genome annotation.}

\item{extUTRlen}{Length to extend 3'UTR to include more near downstream polyA sites.}

\item{DEm6Apval}{cutoff to filter DE-m6A, default is 0.05.}

\item{PDminDist}{Default is 50. Require the distance between  proximal and distal polyA site to be > minDist and < maxDist.}

\item{PDmaxDist}{Default is 1000. Require the distance between  proximal and distal polyA site to be > minDist and < maxDist.}

\item{PDminRatio}{Default is 0.05. If >0 then will filter out polyA sites that do not meet the requirements based on the ratio first, and then filter proximal and distal sites.}

\item{PDfixDistal}{Default is FALSE. True to fix the farthest PA, that is, only look for the proximal site.}

\item{DEAPApval}{cutoff to filter DE-APA, default is 0.05.}
}
\value{
A data frame with each row a gene.
}
\description{
QBGLMmultiDs is a wrapper function of QBGLM for multiple samples, using APA from `QAPA2GLMds` and m6A `m6A2GLMds`.
It first combines multiple APAds and multiple m6Ads. And then links APAds and m6Ads to get per gene data.
And then do QBGLM to test each gene and get betas.
}
\seealso{
Other QBGLM functions:
\code{\link[=FitModel]{FitModel()}},
\code{\link[=QBGLM]{QBGLM()}},
\code{\link[=addPost2FitModels]{addPost2FitModels()}},
\code{\link[=fitQuasiGLM]{fitQuasiGLM()}},
\code{\link[=getM6AperGene]{getM6AperGene()}},
\code{\link[=getRUDperGene]{getRUDperGene()}},
\code{\link[=removeFitErrorModels]{removeFitErrorModels()}},
\code{\link[=testBetas]{testBetas()}}
}
\concept{QBGLM functions}
