% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{removeFitErrorModels}
\alias{removeFitErrorModels}
\title{Remove fitError models in a model list}
\usage{
removeFitErrorModels(models, verbose = TRUE)
}
\arguments{
\item{models}{A model list after fitting QBGLM.}

\item{verbose}{TRUE to show message.}
}
\value{
A list with fitError model removed.
}
\description{
This function removes fitError models in a model list.
}
\seealso{
Other QBGLM functions:
\code{\link[=FitModel]{FitModel()}},
\code{\link[=QBGLM]{QBGLM()}},
\code{\link[=QBGLMmultiDs]{QBGLMmultiDs()}},
\code{\link[=addPost2FitModels]{addPost2FitModels()}},
\code{\link[=fitQuasiGLM]{fitQuasiGLM()}},
\code{\link[=getM6AperGene]{getM6AperGene()}},
\code{\link[=getRUDperGene]{getRUDperGene()}},
\code{\link[=testBetas]{testBetas()}}
}
\concept{QBGLM functions}
