% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{setPACdsSmpInfo}
\alias{setPACdsSmpInfo}
\title{Set sample information for a PACdataset}
\usage{
setPACdsSmpInfo(d, smpInfo)
}
\arguments{
\item{d}{A PACdataset object.}

\item{smpInfo}{A data.frame containing old, new, and group columns.}
}
\value{
A PACdataset object.
}
\description{
Set sample information for a PACdataset
}
