% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{statFitRes}
\alias{statFitRes}
\title{Get summary of the GLM-fitted results}
\usage{
statFitRes(resMp, pval1 = 0.05, qval1 = 0.2, padj1 = 0.1)
}
\arguments{
\item{resMp}{A data frame from `testBetas`, with rawPval1, pval1 and (qval1, padj1...).}

\item{pval1}{cutoff for pval1, default is 0.05.}

\item{qval1}{cutoff for qval1, default is 0.2.}

\item{padj1}{cutoff for qadj1, default is 0.1.}
}
\value{
NULL, print statistics on the screen.
}
\description{
This function count number of genes of statistical significance, with pvalue or adjusted pvalue or qvalue < cutoff.
}
\seealso{
Other output functions:
\code{\link[=getFitDataByGenes]{getFitDataByGenes()}},
\code{\link[=subsetFitRes]{subsetFitRes()}}
}
\concept{output functions}
