% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{subsetFitRes}
\alias{subsetFitRes}
\title{Subset GLM-fitted results}
\usage{
subsetFitRes(resMp, rawPval1 = 1.1, pval1 = 1.1, qval1 = 1.1, padj1 = 1.1)
}
\arguments{
\item{resMp}{A data frame from `testBetas`, with rawPval1, pval1 and (qval1, padj1...).}

\item{rawPval1}{cutoff for rawPval1, default is 1.1 (not filtering).}

\item{pval1}{cutoff for pval1, default is 1.1 (not filtering).}

\item{qval1}{cutoff for qval1, default is 1.1 (not filtering).}

\item{padj1}{cutoff for qadj1, default is 1.1 (not filtering).}
}
\value{
A data frame.
}
\description{
This function subsets genes of statistical significance, with pvalue or adjusted pvalue or qvalue < cutoff, and simplifies columns.
}
\seealso{
Other output functions:
\code{\link[=getFitDataByGenes]{getFitDataByGenes()}},
\code{\link[=statFitRes]{statFitRes()}}
}
\concept{output functions}
