% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6APAreg_funclib.R
\name{testBetas}
\alias{testBetas}
\title{Test the significance of beta for a fitted model list.}
\usage{
testBetas(models, qval = TRUE, FDR = FALSE, empirical = FALSE)
}
\arguments{
\item{models}{A model list after calling fitQuasiGLM and addPost2FitModels.}

\item{qval}{TRUE to adjust the pvalue by qvalue.}

\item{FDR}{TRUE to adjust the pvalue by BH adjustment, default is FALSE.}

\item{empirical}{TRUE to adjust the pvalue by empirical pvalue adjustment of locfdr, default is FALSE.}
}
\value{
A data frame with columns: b0, b1, stdError0, stdError1, df, rawPval0, rawPval1,  tstat0, tstat1, pval0, pval1, and [qval0, qval1, padj0, padj1, emp_pval0, emp_padj0, ..].
}
\description{
This function tests the significance of beta for a fitted model list.
}
\seealso{
Other QBGLM functions:
\code{\link[=FitModel]{FitModel()}},
\code{\link[=QBGLM]{QBGLM()}},
\code{\link[=QBGLMmultiDs]{QBGLMmultiDs()}},
\code{\link[=addPost2FitModels]{addPost2FitModels()}},
\code{\link[=fitQuasiGLM]{fitQuasiGLM()}},
\code{\link[=getM6AperGene]{getM6AperGene()}},
\code{\link[=getRUDperGene]{getRUDperGene()}},
\code{\link[=removeFitErrorModels]{removeFitErrorModels()}}
}
\concept{QBGLM functions}
